added basic modal
